seperate S3 test and visa seed
